+++ src/Modules/Admin/Widgets/Navs/SectionSubmenu.php
@@ -32,6 +32,8 @@ class SectionSubmenu extends Submenu
             entries: $this->getEntriesItem(),
         );
 
+        $this->backUrl ??= $this->getAdminBackUrl($this->model);
+
         parent::configure();
     }